Fix sign in stage 4
The code for waiting on a button press is 0xf9.
The sign towards the end of stage 4 reads:

    "New Magic School"
    We've opened up a
    new magic school!

    <f9>

    Anyone interested
    may apply right now
    for free!

    <9f>

    Principal

              Medusa

The 0x9f-byte is a typo and has been corrected to be 0xf9.
Now the last messages don't fly by and a glitched character in front of
the word Principal is removed.

Fixes #1